move deprecation interfaces from terafoundation to teraslice
