+++ packages/types/src/terafoundation.ts
@@ -26,59 +26,6 @@ export interface SchemaObj<T = any> {
     [key: string]: any;
 }
 
-/** A deprecated field on an asset API */
-export interface AssetAPIDeprecationReason {
-    /** the api name the deprecated field came from */
-    api_name: string;
-    /** the specific field that is deprecated */
-    field: string;
-    description: string;
-}
-
-/** A deprecated field on an asset operation */
-export interface AssetOperationDeprecationReason {
-    /** the op name the deprecated field came from */
-    _op: string;
-    /** the specific field that is deprecated */
-    field: string;
-    description: string;
-}
-
-/** A deprecated field on a job property */
-export interface JobPropertyDeprecationReason {
-    /** the specific field that is deprecated */
-    field: string;
-    description: string;
-}
-
-export interface AssetAPIValidationReason {
-    type: 'assetAPIProperty';
-    kind: 'deprecation';
-    reason: AssetAPIDeprecationReason;
-}
-
-export interface AssetOperationValidationReason {
-    type: 'assetOperation';
-    kind: 'deprecation';
-    reason: AssetOperationDeprecationReason;
-}
-
-export interface JobPropertyValidationReason {
-    type: 'jobProperty';
-    kind: 'deprecation';
-    reason: JobPropertyDeprecationReason;
-}
-
-export type JobValidationReason
-    = | AssetAPIValidationReason
-        | AssetOperationValidationReason
-        | JobPropertyValidationReason;
-
-export interface JobWarning {
-    type: 'JobValidation';
-    reason: JobValidationReason;
-}
-
 export type Schema<T> = {
     [P in keyof T]: Schema<T[P]> | SchemaObj<T[P]>;
 };
